reorganize command scheduling tests
