+++ Domain.Tests/Domain.Tests.csproj
@@ -128,9 +128,9 @@
     <Compile Include="AuthorizationTests.cs" />
     <Compile Include="BloomFilterTests.cs" />
     <Compile Include="CommandContextTests.cs" />
-    <Compile Include="CommandSchedulerMessageExchangeTests.cs" />
+    <Compile Include="CommandSchedulerPipelineTests.cs" />
     <Compile Include="CommandSchedulerTestAggregate.cs" />
-    <Compile Include="CommandSchedulingTests.cs" />
+    <Compile Include="EventSourcedAggregateCommandSchedulingTests.cs" />
     <Compile Include="AggregateCommandTests.cs" />
     <Compile Include="ConfigurationContextTests.cs" />
     <Compile Include="ConfigurationTests.cs" />
@@ -153,6 +153,7 @@
     <Compile Include="PrimitiveConverterTests.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="ReservationServiceTests.cs" />
+    <Compile Include="ScheduledCommandTests.cs" />
     <Compile Include="SerializerTests.cs" />
     <Compile Include="Infrastructure\TestConsequenter.cs" />
     <Compile Include="SnapshotRepositoryExtensionsTests.cs" />
